Incorrect requirements for Java and incorrect dependency version.
Currently, Java 7 is required. In addition, the <version> specified within the <dependency> should be 1.0.1. Background:

tharakd@: "Regarding your concern whether the library compiles with Java 6 or not, I was able to dig up some TT which has the following information and validates my suspicion that the code wont compile with Java 6. https://tt.amazon.com/0074877556"